# pcshopbackend/pcshopbackend/Models/Prebuild.cs
using Microsoft.Extensions.Hosting;
using System.ComponentModel.DataAnnotations;
using System.ComponentModel.DataAnnotations.Schema;

namespace pcshopbackend.Models
{
    public class Prebuild
    {
        public int Id { get; set; }

        [Required]
        public string Name { get; set; }

        [DataType(DataType.Currency)]
        public decimal Price { get; set; }

        [Required]
        public string Description { get; set; }


       public ICollection<Part> parts { get; set; } = new List<Part>();


    }
}